[femutils] Move internal interface in 'internal' subdirectory.
'IDoFLinearSystemImpl' and 'CsrDoFLinearSystemImpl' are internal to
arcanefem and should not be used outside. The corresponding header files
will not be installed with a version of arcanefem.
