+++ femutils/CsrFormatMatrixView.h
@@ -207,12 +207,6 @@ class CsrFormatMatrixView
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
-//! Old name to keep compatibility with existing code.
-using CSRFormatView = CsrFormatMatrixView;
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
 } // namespace Arcane::FemUtils
 
 /*---------------------------------------------------------------------------*/
